BAEL-1376: Varargs Code Example
Moving the "logic" into a test and removing the unnecessary
class.
